Reintegrate Log4j
Modified MagicStrings and evidenceserver.properties to configure the
log4j.xml without  modifications of code. The destination can be
specified by a path which is setable in the properties. The usage of
log4j can be costumized by modifying the log4j xml file.

Hopefully it will be as easy as it seems.